chore(vite): use typescript for configs
