+++ extensions/kube-context/vite.config.ts
@@ -15,10 +15,9 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
-
-import { join } from 'path';
-import { builtinModules } from 'module';
-import { node } from '../../.electron-vendors.cache.json';
+import { join } from 'node:path';
+import { builtinModules } from 'node:module';
+import { defineConfig } from 'vite';
 
 const PACKAGE_ROOT = __dirname;
 
@@ -26,21 +25,21 @@ const PACKAGE_ROOT = __dirname;
  * @type {import('vite').UserConfig}
  * @see https://vitejs.dev/config/
  */
-const config = {
-  mode: process.env.MODE,
+export default defineConfig({
+  mode: process.env['MODE'] ?? 'development',
   root: PACKAGE_ROOT,
   envDir: process.cwd(),
   resolve: {
     alias: {
-      '/@/': join(PACKAGE_ROOT, 'src', '/'),
+      '/@/': join(PACKAGE_ROOT, 'src') + '/',
     },
   },
   build: {
     sourcemap: 'inline',
-    target: `node${node}`,
+    target: 'esnext',
     outDir: 'dist',
     assetsDir: '.',
-    minify: process.env.MODE === 'production' ? 'esbuild' : false,
+    minify: process.env['MODE'] === 'production' ? 'esbuild' : false,
     lib: {
       entry: 'src/extension.ts',
       formats: ['cjs'],
@@ -62,6 +61,4 @@ const config = {
       '@podman-desktop/api': join(PACKAGE_ROOT, '..', '..', '__mocks__/@podman-desktop/api.js'),
     },
   },
-};
-
-export default config;
+});
