+++ packages/renderer/tsconfig.json
@@ -31,6 +31,7 @@
     "types/**/*.d.ts",
     "../../types/**/*.d.ts",
     "../preload/exposedInMainWorld.d.ts",
-    "../api/src/**/*.ts"
+    "../api/src/**/*.ts",
+    "vite.tests.setup.ts"
   ]
 }
