+++ tests/playwright/vite.config.ts
@@ -1,5 +1,5 @@
 /**********************************************************************
- * Copyright (C) 2023-2025 Red Hat, Inc.
+ * Copyright (C) 2023-2024 Red Hat, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,58 +16,46 @@
  * SPDX-License-Identifier: Apache-2.0
  ***********************************************************************/
 
-import { chrome } from '../../.electron-vendors.cache.json';
-import { join } from 'path';
-import { builtinModules } from 'module';
+import { join } from 'node:path';
+import { builtinModules } from 'node:module';
+import typescript from '@rollup/plugin-typescript';
+import { defineConfig } from 'vite';
 
 const PACKAGE_ROOT = __dirname;
-const PACKAGE_NAME = 'preload-docker-extension';
+const PACKAGE_NAME = '@podman-desktop/tests-playwright';
 
 /**
  * @type {import('vite').UserConfig}
  * @see https://vitejs.dev/config/
  */
-const config = {
-  mode: process.env.MODE,
+export default defineConfig({
+  mode: process.env['MODE'] ?? 'development',
   root: PACKAGE_ROOT,
   envDir: process.cwd(),
   resolve: {
     alias: {
       '/@/': join(PACKAGE_ROOT, 'src') + '/',
-      '/@api/': join(PACKAGE_ROOT, '../api/src') + '/',
     },
   },
-  /*plugins: [
-     commonjs({
-       dynamicRequireTargets: [
-         // include using a glob pattern (either a string or an array of strings)
-         'node_modules/ssh2/lib/protocol/crypto/poly1305.js',
-       ]
-       }),
-   ],*/
+  plugins: [],
   build: {
-    sourcemap: 'inline',
-    target: `chrome${chrome}`,
+    sourcemap: true,
+    target: 'esnext',
     outDir: 'dist',
     assetsDir: '.',
-    minify: process.env.MODE !== 'development',
     lib: {
       entry: 'src/index.ts',
-      formats: ['cjs'],
+      formats: ['es'],
+      name: PACKAGE_NAME,
     },
+    // emptyOutDir: true,
+    reportCompressedSize: false,
     rollupOptions: {
-      external: ['electron', ...builtinModules.flatMap(p => [p, `node:${p}`])],
+      plugins: [typescript()],
+      external: ['electron', '@playwright/test', ...builtinModules.flatMap(p => [p, `node:${p}`])],
       output: {
-        entryFileNames: '[name].cjs',
+        entryFileNames: '[name].js',
       },
     },
-    emptyOutDir: true,
-    reportCompressedSize: false,
   },
-  test: {
-    environment: 'jsdom',
-    include: ['src/**/*.{test,spec}.?(c|m)[jt]s?(x)'],
-  },
-};
-
-export default config;
+});
