+++ tsconfig.json
@@ -1,0 +1,14 @@
+{
+  "extends": "@tsconfig/strictest/tsconfig.json",
+  "compilerOptions": {
+    "lib": ["es2023"],
+    "module": "preserve",
+    "target": "es2023",
+    "types": ["node","vitest/globals"],
+  },
+  "include": [
+    "**/vitest.config.ts",
+    "**/vite.config.ts",
+    "vitest.config.ts"
+  ]
+}
