+++ website/package.json
@@ -55,7 +55,8 @@
     "typedoc": "^0.28.14",
     "typedoc-plugin-markdown": "^4.9.0",
     "typescript": "^5.9.3",
-    "vitest": "^3.2.4"
+    "vitest": "^3.2.4",
+    "vite": "^7.2.2"
   },
   "browserslist": {
     "production": [
